requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
py-natpmp==0.2.5
requests==2.31.0
transmission-rpc==7.0.3
typing_extensions==4.9.0
urllib3==2.1.0
